Added pCloud support
Added pCloud support using the official javascript sdk.

To authorize the plugin, the user needs to create an own App on https://docs.pcloud.com/my_apps/ and provide the client ID and secret. These will be used to grant access to the users pCloud drive and to obtain a OAuth access token.